login success authenticate success

+++ src/main/java/com/example/crud/controller/CartItemController.java
@@ -5,10 +5,7 @@ import com.example.crud.entity.Cart;
 import com.example.crud.entity.CartItem;
 import com.example.crud.entity.Product;
 import com.example.crud.entity.User;
-import com.example.crud.service.CartItemService;
-import com.example.crud.service.CartService;
-import com.example.crud.service.ProductService;
-import com.example.crud.service.UserService;
+import com.example.crud.service.*;
 import org.json.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -16,6 +13,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 import java.util.Optional;
 
@@ -28,41 +26,46 @@ public class CartItemController {
     private UserService userService;
     private ProductService productService;
     private CartService cartService;
+    private JwtService jwtService;
 
-    public CartItemController(CartItemService cartItemService, UserService userService, ProductService productService, CartService cartService){
+    public CartItemController(CartItemService cartItemService, UserService userService, ProductService productService, CartService cartService, JwtService jwtService){
         this.cartItemService= cartItemService;
         this.userService= userService;
         this.productService= productService;
         this.cartService= cartService;
+        this.jwtService= jwtService;
     }
 
     //thêm sản phâm vào giỏ hàng
     @PostMapping("/userPage/cartItems/{product-id}")
     public ResponseEntity<CartItem> addProduct(@PathVariable(name = "product-id") long productId,
-                                               @RequestParam(name = "user-id") long userId){
+                                               HttpServletRequest request){
         // tạm thời truyền userId vào, sau lấy userId qua token nên chắc chắn user có tồn tại, không cần check
         try{
-            User user= userService.findById(userId);
-            Cart cart= cartService.getCartByUserId(userId);
-            List<CartItem> cartItemList= cartItemService.getListCartItemInCart(userId);
-            Product product= productService.findById(productId);
-            CartItem cartItemTarget= null;
-            if(cartItemList.size()>0){
-                for(CartItem index: cartItemList) {
-                    if (product.getId() == index.getProduct().getId()) {
-                        cartItemTarget= index;
+            if(jwtService.isCustomer(request)){
+                long userId= jwtService.getCurrentUser(request).getUserId();
+                Cart cart= cartService.getCartByUserId(userId);
+                List<CartItem> cartItemList= cartItemService.getListCartItemInCart(userId);
+                Product product= productService.findById(productId);
+                CartItem cartItemTarget= null;
+                if(cartItemList.size()>0){
+                    for(CartItem index: cartItemList) {
+                        if (product.getId() == index.getProduct().getId()) {
+                            cartItemTarget= index;
+                        }
                     }
                 }
+                if(cartItemTarget!= null){
+                    cartItemTarget.setQuantity(cartItemTarget.getQuantity() + 1);
+                    cartItemService.save(cartItemTarget);
+                } else if(cartItemTarget== null){
+                    cartItemTarget= new CartItem(cart, product, 1);
+                    cartItemList.add(cartItemTarget);
+                    cartItemService.save(cartItemTarget);
+                }
+                return new ResponseEntity(cartItemTarget, HttpStatus.OK);
             }
-            if(cartItemTarget!= null){
-                cartItemTarget.setQuantity(cartItemTarget.getQuantity() + 1);
-                cartItemService.save(cartItemTarget);
-            } else if(cartItemTarget== null){
-                cartItemTarget= new CartItem(cart, product, 1);
-                cartItemList.add(cartItemTarget);
-                cartItemService.save(cartItemTarget);
-            }
-            return new ResponseEntity(cartItemTarget, HttpStatus.OK);
+            return new ResponseEntity<>(HttpStatus.METHOD_NOT_ALLOWED);
         }
         catch(Exception e){
             logger.error(String.valueOf(e));
@@ -72,30 +75,31 @@ public class CartItemController {
     }
 
     //Thêm số lượng sản phẩm trong giỏ hàng
-    @PutMapping("/userPage/cartItems/{product-id}")
-    public ResponseEntity<CartItem> updateQuantity(@PathVariable(name = "product-id") long productId,
-                                                   @RequestParam(name = "user-id") long userId,
-                                                   @RequestBody String data)  {
+    @PutMapping("/userPage/cartItems/{cart-item-id}")
+    public ResponseEntity<CartItem> updateQuantity(@PathVariable(name = "cart-item-id") long cartItemId,
+                                                   @RequestBody String data, HttpServletRequest request)  {
         try{
-            JSONObject jsonObject= new JSONObject(data);
-            int quantity= jsonObject.getInt(InputParam.QUANTITY);
-            User user= userService.findById(userId);
-            Cart cart= cartService.getCartByUserId(userId);
-            Product product= productService.findById(productId);
-            List<CartItem> cartItemList= cartItemService.getListCartItemInCart(userId);
-            for (CartItem cartItem: cartItemList){
-                if(cartItem.getProduct().getId()== productId){
-                    cartItem.setQuantity(quantity);
-                    if(quantity!=0){
-                        cartItemService.save(cartItem);
+            if(jwtService.isCustomer(request)) {
+                JSONObject jsonObject = new JSONObject(data);
+                int quantity = jsonObject.getInt(InputParam.QUANTITY);
+                long userId= jwtService.getCurrentUser(request).getUserId();
+                Cart cart = cartService.getCartByUserId(userId);
+                CartItem cartItem= null;
+                List<CartItem> cartItemList = cartItemService.getListCartItemInCart(userId);
+                for (CartItem index : cartItemList) {
+                    if(index.getCartItemId()== cartItemId){
+                        index.setQuantity(quantity);
+                        cartItem= index;
+                        if (quantity > 0) {
+                            cartItemService.save(cartItem);
+                        } else cartItemService.deleteCartItem(cartItem);
+                        return new ResponseEntity<>(HttpStatus.OK);
                     }
-                    else cartItemService.deleteCartItem(cartItem);
-                    return new ResponseEntity<>(HttpStatus.OK);
                 }
+                cartItemService.save(cartItem);
+                return new ResponseEntity<>(cartItem, HttpStatus.OK);
             }
-            CartItem cartItem= new CartItem(cart, product, quantity);
-            cartItemService.save(cartItem);
-            return new ResponseEntity<>(cartItem, HttpStatus.OK);
+            return new ResponseEntity("Login before processing", HttpStatus.BAD_REQUEST);
         }
         catch (Exception e){
             logger.error(String.valueOf(e));
@@ -105,21 +109,25 @@ public class CartItemController {
 
     //lấy danh sách sản phẩm cùng số lượng trong giỏ hàng
     @GetMapping(value = "/userPage/cartItems")
-    public ResponseEntity<Cart> getListProduct(@RequestParam("user-id") long userId){
+    public ResponseEntity<Cart> getListProduct(HttpServletRequest request){
         try {
-            Cart cart= cartService.getCartByUserId(userId);
-            double amount= 0;
-            List<CartItem> cartItems= cartItemService.getListCartItemInCart(userId);
-            if(cartItems!= null && cartItems.size()>0) {
-                for(CartItem cartItem: cartItems){
-                    amount+= cartItem.getQuantity()* cartItem.getProduct().getPrice();
-                    System.out.println(cartItem.getProduct().getId()+ "   "+ cartItem.getProduct().getName());
+            if(jwtService.isCustomer(request)){
+                long userId= jwtService.getCurrentUser(request).getUserId();
+                Cart cart= cartService.getCartByUserId(userId);
+                double amount= 0;
+                List<CartItem> cartItems= cartItemService.getListCartItemInCart(userId);
+                if(cartItems!= null && cartItems.size()>0) {
+                    for(CartItem cartItem: cartItems){
+                        amount+= cartItem.getQuantity()* cartItem.getProduct().getPrice();
+                        System.out.println(cartItem.getProduct().getId()+ "   "+ cartItem.getProduct().getName());
+                    }
+                    cart.setTotalMoney(amount);
+                    cartService.save(cart);
+                    return new ResponseEntity(cartItems, HttpStatus.OK);
                 }
-                cart.setTotalMoney(amount);
-                cartService.save(cart);
-                return new ResponseEntity(cartItems, HttpStatus.OK);
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
             }
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            return new ResponseEntity("Login before processing", HttpStatus.BAD_REQUEST);
         }
         catch (Exception e){
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -128,10 +136,13 @@ public class CartItemController {
 
 
     @DeleteMapping("/userPage/cartItems")
-    public ResponseEntity<CartItem> removeAllProduct(@RequestParam(name = "user-id") long userId){
-        User user= userService.findById(userId);
-        cartItemService.deleteAllCartItem(userId);
-        return new ResponseEntity<>(HttpStatus.OK);
+    public ResponseEntity<CartItem> removeAllProduct(HttpServletRequest request){
+        if(jwtService.isCustomer(request)){
+            long userId= jwtService.getCurrentUser(request).getUserId();
+            cartItemService.deleteAllCartItem(userId);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        return new ResponseEntity("Login before processing", HttpStatus.BAD_REQUEST);
     }
 
     @GetMapping(value = "/userPage/cartItems/{id}")
